test: eliminate warning re unsafe yaml.load()
Recent pyyaml versions start to warn about using yaml.load(), which is
considered unsafe for an arbitrary input (see CVE-2017-18342). Our input
is controllable (because we wrote tests ourself), but anyway the warning
may be annoying.

The main reason for me, however, is that using of yaml.load() is
disabled for dev-python/pyyaml package on Gentoo. See [1] for more
information.

We did the similar changed for tarantool-python, test-run and tarantool
tests, see [2], [3], [4], [5].

[1]: https://gitweb.gentoo.org/repo/gentoo.git/commit/?id=79ba924d94cb0cf8559565178414c2a1d687b90c
[2]: https://github.com/tarantool/tarantool-python/commit/350771d240a18eec188a53e8c696028b41baa13f
[3]: https://github.com/tarantool/test-run/commit/38400e91c600677fb661154d00459d660fa9880d
[4]: https://github.com/tarantool/test-run/commit/89808d60eb3b5130e227fc1a7866f2ad5a197bea
[5]: https://github.com/tarantool/tarantool/commit/d5fdc533990620a8af520dd68f697b26d9513f7c