test: fix connection failure test for ipv4 only env
An order in which getaddrinfo() returns addresses to try to connect may
vary from system to system. In php:7.0-cli based docker container
127.0.0.1 address is tried first, then ::1.

The attempt to connect the IPv4 address gives errno=111 (ECONNREFUSED,
'Connection refused') and it would be correctly matched by the test.

An attempt to connect the IPv6 address gives errno=99 (EADDRNOTAVAIL,
'Cannot assign requested address'), which is reported in the exception
message and considered as the failure by the test case.

IPv4 is likely supported everywhere, so restrict the attempts to the
IPv4 address.