test: fix failed name resolution test for some envs
Under certain network conditions it is possible that getaddrinfo() will
return EAI_AGAIN rather than EAI_NONAME for an unknown host.

It can be easily reproduced when a network interface is down, however it
is not the only case when EAI_AGAIN appears. It seems some DNS servers
may report a temporary failure in case of an unknown hostname (at least
it looks so, I am unable re-check now).