test: allow to verify all configurations locally
Provide the script, which run the whole test suite for the connector on
all supported tarantool and php versions. It is useful to verify changes
locally.

The script uses php:<version>-cli docker images and the ./test.sh script
inside (exactly what is called on Travis-CI). The ./test.sh script is
modified to work under those images:

* install necessary packages (gnupg2, lsb-release, apt-transport-https),
  when when they're not installed
* support tarantool repositories for Debian (not only Ubuntu)
* don't require 'sudo' when the current user is root (usual for docker)
* don't require 'pip', install python-yaml from packages (pip is not
  available by default in php's images)
* while we're here, changed GPG key and repository URLs from http:// to
  https:// (it is redirected to https:// anyway)
* fix tarantool-1.6 installation on Debian Stretch

Downside: APT repositories updating takes significant time and it is
performed for each configuration.