Properly initialize Tarantool instance properties
The proper initialization is important for correct work of an inherited
class. I cannot provide exact description what was broken and becomes
right, but at least the test case with inherited class with a property
starts to work correctly. Some online materials strongly suggest to call
object_properties_init() function in a create_object handler, see [1],
[2], [3].

The problem was catched at the existing test suite on mocking tests when
phpunit was updated to 6.5.14 (now it is 4.8.24) and appropriate changes
were made for the tests (say, using of createMock() instead of
getMock()). So this commit is prerequisite to run the test suite on more
fresh phpunit, which is necessary to test the connector on php-7.1+. The
changes for php-7.1+ will land in future commits.

It seems the problem was introduced in [4].

[1]: https://wiki.php.net/internals/engine/objects
[2]: http://www.phpinternalsbook.com/php5/classes_objects/custom_object_storage.html
[3]: https://phabricator.wikimedia.org/T59292
[4]: 9f5a2828de9f49ceb07b9e803424da9b27802700 ('updated PHP7 implementation')

Fixes #135.