travis-ci: install appropriate phpunit version
We should use different phpunit versions on different php versions (see
[1]). Since we going to enable php from 7.0 to 7.4, the logic of
test-run.py is changed to find `phpunit` command in PATH directories,
but the Travis-CI script (test.sh) installs appropriate phpunit version
to /usr/local/bin.

Note: On the first glance it looks that we can skip phpunit installing
on Travis-CI, because it is provided already. However it ships
phpunit-7.5.0 for php-7.0 environment, but it works only on php-7.1+.

Note: We can not just save downloaded phpunit into /usr/local/bin,
because Travis CI have a directory with its own phpunit executable in
PATH prior standard directories. So we create our own directory and add
it to PATH at beginning.

Tests are changed to be compatible with phpunit-6+, see [2] and [3].

Aside of this, removed some dead code from test-run.py.

The next commit will remove test/phpunit.phar file, it is not used
anymore.

[1]: https://phpunit.de/supported-versions.html
[2]: https://thephp.cc/news/2017/02/migrating-to-phpunit-6
[3]: https://github.com/sebastianbergmann/phpunit/pull/2898