test: support phpunit-7
Added the compatibility layer for phpunit-6 and phpunit-7.

The problem, which is solved here, is that PHP 7.0 does not support
`void` return value declaration, but phpunit-7 (which supports
php-7.[1-3]) requires it for some methods (see [1]). The compatibility
layer is necessary to run the same testing code on different php/phpunit
versions.

The implementation idea is borrowed from Symfony's [PHPUnit Bridge][2].
There is also broad discussion about this problem in the Drupal's
[issue 3063887][3].

In fact the test suite works on phpunit-8 too, but phpunit produces
several warnings about using of deprecated functionality. It will be
fixed in the following commits.

[1]: https://phpunit.de/announcements/phpunit-7.html
[2]: https://symfony.com/doc/current/components/phpunit_bridge.html#removing-the-void-return-type
[3]: https://www.drupal.org/project/drupal/issues/3063887