test: eliminate phpunit-8 deprecation warnings
After enabling phpunit-7, phpunit-8 starts to work too, but gives
several deprecation warnings:

* Using of `assertContains()` for search in a string.
* Using of `@expectedException` and related annotations.

Functionality that is deprecated in phpunit-8 will be disabled in
phpunit-9. Since we plan to use phpunit-9 for testing on php-7.3 and
php-7.4, it worth to fix the warnings now.

This commit updates the test suite to use the new constructions and
provides the compatibility layer for phpunit-6, which does not contain
`assertStringContainsString()` method.

The commit also disables caching of testing results, which was enabled
by default in phpunit-8 (see [1]). The test suite is quite small and the
feature does not add any value. The reason for disabling is to protect
ourself from side effects of this caching: to be honest I don't know how
it may change testing results.

[1]: https://phpunit.de/announcements/phpunit-8.html