Raise exception for update by an absent field name
It seems the exception "No field '<...>' defined" was not raised since
introducing of named fields support in
31cf8a8b4c8cec99fd6e03a80b2df42d9b3cb27d ('Next version of tarantool-php
for php7').

There is the test for this case
(DMLTest::test_08_update_error_nosuchfield), but it accepts any error
message, because of a typo in an annotation name. Fixed this typo.

The problem was found by this test when the annotation was replaced by a
call in order to eliminate a phpunit-8 warning and to support phpunit-9.
This change will land in a next commit.