travis-ci: verify CentOS / Fedora packages
Verify `packpack` based builds for CentOS and Fedora distributions using
Travis CI. Deployment is not enabled here: it is subject for further
commits.

The key point is that we'll build the extension for php version that is
provided by an OS itself.

Why don't build against PHP versions from widely used Remi's repository?
There are two reasons:

* It will introduce dependency on a third-party repository for our
  third-party repository, at least in context of php-tarantool RPM
  packages. I guess it would quite unobvious from a user perspective.
* We don't control when PHP ABI version will be changed in Remi's
  repositories. Is there a politic about this? I don't aware of it.
  We can get non-working packages after an update of PHP version in
  Remi's repo.

So the decision for now is to provide packages for a PHP version
provided by a distribution itself (from its base repositories).

I plan to contact Remi and consult how to handle everything in a way
that would be convenient for a user. Maybe it would be good to provide
sources, spec and build tools from our side and ask Remi to include the
extension into his repositories.

Part of #117