build: allow to test RPM / Deb packages locally
Provided the script to build a RPM / Deb package, install it inside a
docker container based on an appropriate image and verify that the main
class of the extension (Tarantool) is available from the PHP
interpreter.

The script builds a package just like it is performed in CI: using
packpack. It should be available in PATH directories.

Only RPM packages are here now, but the following commits will enable
Deb packages too.

Part of #117