travis-ci: verify Deb packages for PHP 7.0
Verify Deb packages on distributions, which provides PHP 7.0: it is
Debian Stretch and Ubuntu Xenial.

Now the Debian build files (debian/ directory) allow to build against
PHP 7.0 only (not PHP 7.1+), because of hardcoded package name and
installation paths. That is why newer Debian and Ubuntu distributions
are not enabled here. It will be fixes in the following commits.

Part of #117