test: use strict types mode
It is the rule of thumb to run tests in the strict types mode, because
it may reveal subtle problems. See #154 for example.

Updated the test suite to cast tarantool port (acquired from
PRIMARY_PORT environment variable) from string to int. Aside of this,
raise an error when the port is not set: it'll allow to fail fast and
provide a good diagnostic when a configuration problem occurs.