schema: add support new _index system space format
After tarantool-1.7.5-153-g1651fc9be the new _index format was
introduced. We should support it to fetch a schema from
a tarantool-1.7.5.153+ instance.
When an index parts do not use parameters except name and type
the index info is stored in the old format in _index system space.
When an index part uses is_nullable or collation parameter, then
the new format will be used.

Close #151