PHP 8.0: define TSRMLS* macros as no-op
The TSRM* macros are no-op since PHP 7.0. PHP 8.0 drops them at all.
Define them as no-op in our code when building against PHP 8 and newer
to overcome a build fail. See [1] and [2] (1.c) for details.

Those macros could be removed from the code entirely (both definitions
and usages), since we support PHP 7.0+. It is to be done later.

[1]: https://wiki.php.net/rfc/native-tls
[2]: https://github.com/php/php-src/blob/php-8.0.18RC1/UPGRADING.INTERNALS

Part of #171