readme: get rid of 1.6 in the title
The 'driver for Tarantool 1.6' phrase was written when Tarantool 1.6 is
just released and there was a need to differentiate a driver that
implements Tarantool 1.6+ protocol from one that leans on Tarantool
1.5 protocol (those are not compatible).

Starting from Tarantool 1.6 the protocol evolves in the backward
compatible manner. The driver actually supports any Tarantool versions
since 1.6.

Tarantool 1.5 is not supported for years, so there is no practical sense
to highlight that the driver is for 1.6+. It only confuses, because one
may assume that the driver does not support Tarantool versions newer
than 1.6.

The driver is tested on versions from 1.6 to 2.5 and should work for
newer tarantool versions as well.